requirements.txt
torch
transformers
flask
flask-cors
requests
Pillow
numpy
timm
einops
huggingface_hub

